Completed the League module

+++ league-tracker/src/main/java/com/example/leaguetracker/data/LeagueRepository.java
@@ -2,5 +2,5 @@ package com.example.leaguetracker.data;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface LeagueRepository extends CrudRepository<League, Integer> {
+public interface LeagueRepository extends CrudRepository<League, Long> {
 }
